playback: media-session backends are now direct coordinator sinks
Retire MediaSession, MediaSessionThread, MediaSessionPlaybackSink,
MediaSessionBackend, and g_media_session. Each platform's media-session
integration (MPRIS over D-Bus, MPNowPlayingInfoCenter on macOS,
SystemMediaTransportControls on Windows) is now its own PlaybackEventSink
with its own thread. JS metadata IPCs route through PlaybackCoordinator
via new MetadataChanged / ArtworkChanged / QueueCapsChanged / Seeked
events, so coord is the single ingress for everything backends consume.

CefThreadPlaybackSink renamed to QueuedPlaybackSink (the "Cef" was
incidental — the consumer thread pumping it is just whatever thread polls
wake().fd()). Platform sinks inherit it and pump from their own thread.

MediaMetadata / MediaType / PlaybackState moved from player/media_session.h
to playback/event.h. src/player/ directory removed entirely.

Inbound transport callbacks (Play/Pause/Next/Previous/Seek/SetRate from
MPRIS clients, MPRemoteCommandCenter, SMTC ButtonPressed) are wired
directly into g_mpv / g_web_browser inside the per-platform handlers,
no MediaSession::onPlay-style indirection.
