+++ src/common.h
@@ -10,14 +10,13 @@ extern Color g_video_bg;
 
 #include "platform/platform.h"
 #include "mpv/handle.h"
-#include "player/media_session.h"
+#include "playback/event.h"
 
 class WakeEvent;
 
 extern MpvHandle g_mpv;
 extern Platform g_platform;
 
-class MediaSessionThread;
 class ThemeColor;
 class PlaybackCoordinator;
 extern PlaybackCoordinator* g_playback_coord;
@@ -25,7 +24,6 @@ extern PlaybackCoordinator* g_playback_coord;
 void initiate_shutdown();
 extern std::atomic<bool> g_shutting_down;
 extern WakeEvent g_shutdown_event;
-extern MediaSessionThread* g_media_session;
 extern ThemeColor* g_theme_color;
 
 // Display refresh rate (Hz) — updated from mpv's display-fps property observation.
