+++ src/event_dispatcher.h
@@ -5,8 +5,8 @@
 #include <memory>
 #include <vector>
 
-class CefThreadPlaybackSink;
-class CefThreadActionSink;
+class QueuedPlaybackSink;
+class QueuedActionSink;
 
 // mpv → coordinator bridge. mpv_digest_thread normalizes mpv events and
 // publishes them here; cef_consumer_thread routes every playback-relevant
@@ -19,9 +19,9 @@ void publish(const MpvEvent& ev);
 // Called by run_with_cef before starting the consumer thread so the
 // dispatcher knows which sinks to pump. Sinks must outlive the consumer
 // thread.
-void register_cef_thread_sinks(
-    std::vector<std::shared_ptr<CefThreadPlaybackSink>> event_sinks,
-    std::vector<std::shared_ptr<CefThreadActionSink>> action_sinks);
+void register_queued_sinks(
+    std::vector<std::shared_ptr<QueuedPlaybackSink>> event_sinks,
+    std::vector<std::shared_ptr<QueuedActionSink>> action_sinks);
 
 void cef_consumer_thread();
