+++ src/playback/sinks/idle_inhibit_sink.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include "cef_thread_sinks.h"
+#include "queued_sink.h"
 
 // Watches phase + media_type from ev.snapshot and updates the platform
 // idle inhibit level.
-class IdleInhibitSink final : public CefThreadPlaybackSink {
+class IdleInhibitSink final : public QueuedPlaybackSink {
 protected:
     void deliver(const PlaybackEvent& ev) override;
 };
