+++ tests/playback_coordinator_test.cpp
@@ -4,6 +4,7 @@
 #include "playback/coordinator.h"
 
 #include <chrono>
+#include <functional>
 #include <memory>
 #include <mutex>
 #include <thread>
